wenshu_utils-master/requirements.txt
lxml
pycryptodomex
PyExecJS

# 用于测试
requests